converter/storage: move volume source conversion to storage/volume.go
The volume source helpers only serve the disk conversion loop. Grouping
them in the storage sub-package prepares for extracting the disk loop
into a DiskConfigurator.

toApiReadOnly is exported because Convert_v1_Disk_To_api_Disk, which
stays in converter.go, also calls it.
