+++ pkg/virt-launcher/virtwrap/converter/BUILD.bazel
@@ -11,18 +11,12 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/cloud-init:go_default_library",
-        "//pkg/config:go_default_library",
-        "//pkg/container-disk:go_default_library",
-        "//pkg/emptydisk:go_default_library",
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
-        "//pkg/host-disk:go_default_library",
         "//pkg/os/disk:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/safepath:go_default_library",
         "//pkg/storage/reservation:go_default_library",
         "//pkg/storage/types:go_default_library",
-        "//pkg/storage/volumepath:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
@@ -81,6 +75,7 @@ go_test(
         "//pkg/virt-launcher/virtwrap/converter/arch:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/compute:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/network:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/storage:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/types:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/vcpu:go_default_library",
         "//pkg/virt-launcher/virtwrap/device/hostdevice/generic:go_default_library",
