+++ pkg/virt-launcher/virtwrap/converter/storage/BUILD.bazel
@@ -2,11 +2,22 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["virtiofs.go"],
+    srcs = [
+        "virtiofs.go",
+        "volume.go",
+    ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/storage",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/cloud-init:go_default_library",
+        "//pkg/config:go_default_library",
+        "//pkg/container-disk:go_default_library",
+        "//pkg/emptydisk:go_default_library",
+        "//pkg/host-disk:go_default_library",
+        "//pkg/storage/volumepath:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/types:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/virtio:go_default_library",
         "//pkg/virtiofs:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
     ],
